pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mywardrobe"
version = "0.1.0"
description = "Multimodal recommendation system for wardrobe management"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "readme.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.1",
    "torchvision",
    "transformers",
    "openai-clip",
    "pillow",
    "numpy",
    "faiss-cpu",
    "tqdm",
    "supabase",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-dir]
"" = "src" 